requirements.txt
streamlit
torch
torchvision
Pillow
numpy
opencv-python
pandas
matplotlib
plotly
gdown
random
